# Conflicts:
#	src/Model/Shapes/Shape.java
#	src/Model/Square.java

+++ src/Model/Shapes/Square.java
@@ -8,7 +8,7 @@ public class Square extends Shape{
         return new Square(getX(),getY(),getWidth(),getHeight(),getIsFilled());
     }
 
-    public Square(int x, int y, int width, int height, boolean isFilled){
+    public Square(double x, double y, int width, int height, boolean isFilled){
         super(x,y,width,height,isFilled);
     }
 }
